+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## ?.?.?
+
+- Calendar cards can now be configured to display a limited amount of events.
+
 ## 0.3.0
 
 - Fixed docker build.